# openapi.yaml
openapi: 3.0.3
info:
  title: "My AR Project API"
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: ローカル開発用サーバ
  - url: http://app:8080
    description: Docker内部用サーバ
paths:
  /fireworks/latest-id:
    get:
      summary: "現在登録されている花火の最大ID（削除済みを含む）を取得する"
      operationId: "GetLatestFireworkId"
      responses:
        '200':
          description: "最大ID。花火が1件も存在しない場合は 0"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LatestFireworkIdResponse"
  /fireworks:
    get:
      summary: "花火の一覧を取得する"
      operationId: "GetFireworks"
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
            description: "created_at の開始日 (YYYY-MM-DD、当日含む)"
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
            description: "created_at の終了日 (YYYY-MM-DD、当日含む)"
      responses:
        '200':
          description: "花火の一覧"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FireworkResponse"
    post:
      summary: "花火を作成する"
      operationId: "CreateFirework"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FireworkCreateRequest"
      responses:
        '201':
          description: "花火が作成されました"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FireworkResponse"
  /fireworks/{id}:
    get:
      summary: "IDで指定した花火を取得（isShareable が false の場合は存在しない場合と同様に404を返す）"
      operationId: "GetFireworkById"
      parameters:
        - name: id
          in: path
          required: true
          description: "花火のID"
          schema:
            type: integer
            format: int64
        - name: resolution
          in: query
          required: false
          schema:
            type: string
            enum: [low, mid, high]
            default: low
            description: "取得する解像度（現時点では無視。将来拡張用）"
      responses:
        '200':
          description: "花火の詳細"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FireworkResponse"
    put:
      summary: "IDで指定した花火の共有設定を更新"
      operationId: "UpdateFirework"
      parameters:
        - name: id
          in: path
          required: true
          description: "花火のID"
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FireworkUpdateRequest"
      responses:
        '200':
          description: "花火が更新されました"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FireworkResponse"
    delete:
      summary: "IDで指定した花火を削除"
      operationId: "DeleteFirework"
      parameters:
        - name: id
          in: path
          required: true
          description: "花火のID"
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: "花火が削除されました"
          content: {}
components:
  schemas:
    LatestFireworkIdResponse:
      type: object
      description: "現在登録されている花火の最大ID（削除済みを含む）のレスポンス"
      required:
        - latestId
      properties:
        latestId:
          type: integer
          format: int64
          description: "花火が1件も存在しない場合は 0"
          example: 5
    FireworkCreateRequest:
      type: object
      description: "花火を作成するためのリクエストボディ"
      required:
        - isShareable
        - image
      properties:
        isShareable:
          type: boolean
          description: "花火が共有可能かどうか"
          example: false
        image:
          type: string
          format: binary
          description: "花火の元となる画像ファイル"
    FireworkUpdateRequest:
      type: object
      description: "花火を更新するためのリクエストボディ"
      required:
        - isShareable
      properties:
        isShareable:
          type: boolean
          description: "花火が共有可能かどうか"
          example: true
    FireworkResponse:
      type: object
      description: "花火のレスポンスデータ"
      required:
        - id
        - isShareable
        - imageUrl
      properties:
        id:
          type: integer
          format: int64
          description: "花火のID"
          example: 1
        isShareable:
          type: boolean
          description: "花火が共有可能かどうか"
          default: false
          example: true
        imageUrl:
          type: string
          nullable: true
          description: "元画像の公開 URL。旧レコードは null"
        createdAt:
          type: string
          format: date-time
          description: "花火の作成日時"
          example: "2023-10-01T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: "花火の更新日時"
          example: "2023-10-01T12:00:00Z"